+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added a `step_on_batch` flag to the early stopping callback
 - Added multi image support to `imaging`
 - Added a callback to unpack state into torchbearer.X at sample time for specified keys and update state after the forward pass based on model outputs. This is useful for using DataParallel which pass the main state dict directly. 
+- Added callback for generating confusion matrices with PyCM
 ### Changed
 - Changed history to now just be a list of records
 ### Deprecated